populates the main screen with cards that are drawn from the deck.
uses the AspectVGrid view from the class example code provided. in
the future, this likely need to be updated to enforce a minimum
size for each card, and possibly wrapped in a scroll view.

the Set model now contains cards in the deck as well as cards that
were drawn and visible.