hook up draw 3 more cards and new game buttons to intents
once these buttons were hooked up, it revealed a bug in the
AspectVGrid implementation, which seems to stem from a bug in
LazyVGrid. changes in the width of an adaptive grid column (from
smaller to larger value) would not trigger a reevaluation of the
height of the LazyVGrid. therefore the grid would not necessarily
try to build all the child views (it was acting a little too lazy).

in order to work around this issue, i added the use of the .id()
view modifier to the LazyVGrid. this invalidates the LazyVGrid each
time any item is changed, which coincides with width changes, and
resolves the issue. this workaround might interfere with the
ability to animate cards later on.

another bug was found and fixed: the deck is checked to have cards
before drawing cards from it.