adds card selection and replacing drawn cards from the deck
this change involves quite a bit of code in many parts. the CardView
is updated to use the edge to represent selected cards, cards in a
matching selection, and cards in an unmatching  selection. the model
is updated to represent selected cards in a set of indicies. the model
also has the logic for updating the drawn cards on each selection.

a small additional piece is disabling the Draw 3 Cards button when
the deck is empty.

besides the few TODOs (or bugs) noted, this commit nearly completes
the requirements. the only remaining requirement is the algorithm
for determining if the selection is a matching Set.

there is still some clean up / refactoring that could help, esp in
CardView and AspectVGrid.