Add deck and matched card stacks to the control area
The model is adjusted to store matched cards in an array. That property is exposed through the view model through to the view.

The model logic is adjusted to no longer replace matched cards automatically (task 2).

Cards in the deck are now drawn face down. The deck is interactive and replaces the "Draw 3 More Cards" button.

No animations are implemented and we need to revisit how sizing and layout is handled. Face up cards need a background color (maybe dependant on dark mode).

Task 9 is incomplete.