implements required task 9, to remove cards in the selection when drawing new cards (when that selection contains matched cards).
this makes the ViewModel even less useful, and pushes more logic into the Model. at this point I'm beginning to question the purpose
of having a ViewModel at all. it's only real use at this point is having an object of reference type (as opposed to value type) that my
View can track as the observed object.