Adds  a preview of the interface in the landscape orientation.
This is a checkpoint on the way to adjusting the layout to something that looks better in both portrait and landscape orientations.

At this point, the biggest question is why the `controls` view is sized differently in each orientation? Within the parent VStack, it is the
least flexible child, and so it should be offered about 50% of the height and the entire width. In portrait, controls takes all the offered
width, and part of the offered height. In landscape, controls takes all the offered height and part of the offered width (this leads to the
area outside the safe area not being colored by the FeltGreen background color). Intuitively, it feels like it might have something to
do with the aspect ratio as a constraint when controls tries to determine its own size.