Small cleanup of view code.
Found answers to the layout related questions in the previous commit. As I suspected, the .aspectRatio() modifier plays a role in the
difference between the landscape and portrait orientation layouts. In the portrait orientation, the controls view contains deck and
matchedCards views which are *width-limited*. In this layout, the controls view takes up all the proposed width, and the height is based
on the aspectRatio. In the landscape orientation, the controls view contains deck and matchedCards views which are *height-limited*.
In this case, they take up all the proposed height, and the width is based on the aspectRatio.

A more sensible approach would be to limit the size of the individual cards in the deck and matchedCards views to be no larger than the
size of a card in the playArea. However this would require communicating the geometry of a card view across distant siblings. Based on
some preliminary research, this is best accomplished in SwiftUI using PreferenceKeys.