Implements animations on completing a matching selection or mismatching selection
This makes use of a custom GeometryEffect, which is an animatable view modifier. It uses a technique that imitates "keyframe animation" using the `pct` value between 0 and 1 as the time for the animation. It also uses an implicit animation to separate the animation of the shake from the other animations occurring during a card choose intent. The implicit animation also helps to make the animation "one-way", so it occurs on the transition to a certain state, but not on the transition out of that state.

There is one downside to this technique. The animatable view modifiers inside the CardView are now subject to the implicit animation. I thought that wouldn't be the case since the `.animation(_:value:)` method is given a value of the `cardSelectionState`. However, it might just be that since the CardView's border color and width (animatable view modifiers) also change at the same time as `cardSelectionState`, that this value isn't specific enough to prevent the animation from affecting the CardView. I was hoping for some way of resetting the implicit animation so that it doesn't affect the wrapped view.