Limits the height of the controls to 1/4 the total size of the body or less.
Uses a pretty strange technique to accomplish the size limits. The controls are given first layout priority to allow it to get as much height as it can use in the remaining 1/4 of the height of the VStack. If it takes less, then the sibling ZStack takes up the rest, because the frame has a minHeight, and the ZStack's children can all expand to fit all the available space.

See: https://stackoverflow.com/a/71853238/305340