Adds code for debugging the game.
Set and Set.Card both gain CustomStringConvertible conformance so
that the SetGame view model can print out the state of the game
in key transitions.

In AspectVGrid, a red border is added so that the frame of the
inner LazyVGrid can be seen. This will be helpful for understanding
the layout during animations.

In SetGameView, when the New Game button is pressed, we use the
game printing capability of SetGame to capture the state before
and after animations.