No longer pursuing layout based on relative card sizes
I now understand that a layout based on relative card sizes would lead
to undesirable changes in the height of the `controls` view during
gameplay. Specifically, there are two ways to define the layout based on
relative card sizes: using the size of cards in `controls` as the
minimum size for cards in `playArea` or using the size of cards in
`playArea` as the maximum for the cards in `controls`. The first way
would not constrain the size of cards in `controls` in wider layouts
such as an iPhone in landscape, causing it to take up potentially over
half of the screen area. The second way would allow cards in the
`playArea` to increase in size (as the number of cards becomes fewer)
and then also allow the `controls` area to increase in size. That would
create layout shifts where the height of the `controls` area increased
when there were fewer cards in play and decreased as more cards were
dealt.

The situation above is so undesirable that a the preferred path forward
is a layout where the height of `controls` is limited to 1/4 of the
total screen height, and the `playArea` gets the rest. Unfortunately,
expressing this intended layout is not so simple. The
`.frame(maxHeight:alignment:)` view modifier cannot express this because
the frame view itself (which wraps the modified view) will still use
the height it is offered when the modified view chooses to use less.
After researching the `.frame()` modifier and related built-in layout
techniques in SwiftUI, I learned that the layout intention cannot
be directly expressed, but something close with the same outcome can.
In this implementation, `playArea` is given a minimum height of 3/4 of
the total screen height, and `controls` is allowed to lay itself out
first.

Given this implementation, one additional layout constraint would be
helpful. While the size of `controls` and `playArea` are already
determined, we could still try to make sure that cards in `controls` are
never bigger than cards in `playArea`. This would be done by measuring
the size of cards in `controls` and using the width as the minimum item
width in `playArea` (currently implemented with `AspectVGrid`). This
can potentially be implemented using a custom `LayoutValueKey`. Before
investigating and prototyping this implementation, I'd like to first
take an inventory of remaining animation issues so that it's possible
to identify whether or not a regression occurs.