Fixes layering order in dealing animations
This is implemented by explicitly ordering the cards using a zIndex()
modifier. The depth of each card is calculated using a helper.

The solution is also applied to the discardPile, since it has roughly
the same problem. This allowed us to remove the PileView completely,
since using it actually used more code than not using it. We may add
this back if and when we work on messy looking piles of cards.

Initially using explicit zIndex() ordering seemed unnecessary because
we could simply reverse the collection when we needed the opposite
ordering. It turns out that SwiftUI's matchedGeometryEffect and
animation system does not preserve the implicit order of views. Through
some trial and error we found that as soon as the animation begun, the
top card would appear behind the rest of the deck. The explicit ordering
solved that issue.

Added a task to the list, which is to improve animations where a group
of cards are moving (such as dealing) by making them one-by-one. This
would improve the dealing animation because even now it's not clear that
the top cards (after the first) are being dealt.