Refactor newGame() and dealInitialCards() actions for animation debugging
Removed the calls to groupOfCards() to generate the animation. This helper should eventually be removed.
Added lines of code that can be uncommented to slow down the specific resetting and dealing animations.
Refactored the print() calls to be a little less verbose/repetitive.

The behavior has not changed in any noticeable way in this commit. When the New Game button is pressed, the animations
for reset and the dealing of initial cards do not overlap at all (there is a 0.1s pause in between). However, I used a sample
app to confirm that SwiftUI should be able to correctly "interrupt" an animation even if the two did overlap. There is still some
animation issue with cards that are in the discard pile when the New Game button is pressed and then end up in the play
area after the initial deal. The transition into the play area does not follow the matchedGeometryEffect (they slide in from the right)
and they are timed at the same time as the reset (which only removes cards from the play area and adds them to the deck).
I recorded some thoughts in the README.