Migrated all code to target iOS17 with latest Xcode project settings
One significant implementation change was the adoption of Observable
framework for models, in place of the ObservableObject and @Published
mechanisms used before.

A few unused files were deleted.

The #Preview macro replaces the PreviewProvider implementation

There are still plenty of TODOs, which deserve some more migration
attention.

The app is running with the same functionality as before.